Fix integer size problem for the ColumnValue class

+++ gerrydb_meta/models.py
@@ -548,7 +548,7 @@ class ColumnValue(Base):
     __tablename__ = "column_value"
     __table_args__ = (UniqueConstraint("col_id", "geo_id", "valid_from"),)
 
-    val_id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    val_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     col_id: Mapped[int] = mapped_column(
         Integer,
         ForeignKey("column.col_id"),
